requirements.txt
pyyaml>=6.0
PySide6-Essentials==6.4.0.1